refactor: split settings view overlays
