+++ src/ui/views/settings.rs
@@ -2,6 +2,15 @@
 //! field rows. State lives in `App.settings`; like the library view, the `ListState`
 //! that highlights the focused row is rebuilt fresh each frame from a `usize` index.
 
+mod dialogs;
+mod recording;
+mod spotify;
+
+pub use dialogs::{render_confirm, render_conflict};
+pub use recording::{render_recording_settings, render_recordings_browser};
+pub(super) use spotify::render_spotify_import_mode_dropdown_popup;
+pub use spotify::render_spotify_picker;
+
 use ratatui::Frame;
 use ratatui::layout::{Alignment, Constraint, Layout, Rect};
 use ratatui::style::{Color, Modifier, Style};
@@ -13,9 +22,9 @@ use crate::app::{App, MouseTarget, ScrollSurface, StatusKind};
 use crate::config::{
     FPS_DEFAULT, FPS_MAX, FPS_MIN, SEEK_SECONDS_MAX, SEEK_SECONDS_MIN, SPEED_MAX, SPEED_MIN,
 };
-use crate::keymap::{self, Action, Conflict, KeyContext};
+use crate::keymap::{self, Action, KeyContext};
 use crate::settings::{BAND_GAIN_MAX, BAND_GAIN_MIN};
-use crate::settings::{Field, FieldKind, SettingsConfirm, SettingsState, SettingsTab};
+use crate::settings::{Field, FieldKind, SettingsState, SettingsTab};
 use crate::t;
 use crate::theme::ThemeConfig;
 use crate::theme::ThemeRole as R;
@@ -689,7 +698,7 @@ fn render_fields(frame: &mut Frame, app: &App, st: &SettingsState, area: Rect) {
     // Per-control click targets (checkbox / arrows / button / text) on top of the row rects, so
     // they win where they overlap (`mouse_target_at` takes the last-registered match).
     register_field_controls(app, st, area, offset, &display_to_field);
-    render_spotify_import_mode_dropdown(frame, app, st, area, offset, &display_to_field);
+    spotify::render_spotify_import_mode_dropdown(frame, app, st, area, offset, &display_to_field);
     // Right-border scrollbar tracking the viewport; a no-op when every row fits.
     buttons::render_list_scrollbar(
         frame,
@@ -790,197 +799,6 @@ fn register_field_controls(
     }
 }
 
-fn render_spotify_import_mode_dropdown(
-    frame: &mut Frame,
-    app: &App,
-    st: &SettingsState,
-    area: Rect,
-    offset: usize,
-    display_to_field: &[Option<usize>],
-) {
-    let Some(selected) = st.spotify_import_mode_dropdown else {
-        return;
-    };
-    let fields = st.fields();
-    let Some(field_index) = fields
-        .iter()
-        .position(|field| *field == Field::SpotifyImportMode)
-    else {
-        return;
-    };
-    let Some(display_row) = display_to_field
-        .iter()
-        .position(|mapped| *mapped == Some(field_index))
-    else {
-        return;
-    };
-    if display_row < offset {
-        return;
-    }
-    let visible = display_row - offset;
-    if visible >= area.height as usize {
-        return;
-    }
-
-    let theme = &st.draft.theme;
-    let gutter = buttons::text_width(HL_SYMBOL);
-    let vx = area.x + gutter + other_label_width(st.tab) as u16;
-    let available_width = area.right().saturating_sub(vx);
-    if available_width == 0 {
-        return;
-    }
-    let value = field_value_text(
-        app,
-        st,
-        Field::SpotifyImportMode,
-        field_index == st.row,
-        available_width as usize,
-    );
-    let label_width = buttons::text_width(&value)
-        .max(
-            crate::config::SpotifyImportMode::ALL
-                .iter()
-                .map(|mode| buttons::text_width(mode.label()))
-                .max()
-                .unwrap_or(0)
-                + 4,
-        )
-        .min(available_width);
-    let leading = vx.saturating_sub(area.x) as usize;
-    // Drop below the row; when the list is short (e.g. the docked player bar shrank it)
-    // and the options wouldn't fit, drop up above the row instead of clipping them away.
-    let modes = crate::config::SpotifyImportMode::ALL.len() as u16;
-    let anchor_y = area.y + visible as u16;
-    let below = area.bottom().saturating_sub(anchor_y + 1);
-    let above = anchor_y.saturating_sub(area.y);
-    let start_y = if below >= modes || above < modes {
-        anchor_y + 1
-    } else {
-        anchor_y - modes
-    };
-    for (idx, mode) in crate::config::SpotifyImportMode::ALL
-        .iter()
-        .copied()
-        .enumerate()
-    {
-        let y = start_y + idx as u16;
-        if y >= area.bottom() {
-            break;
-        }
-        let focused = idx == selected;
-        let marker = if focused { ">" } else { " " };
-        let text = format!(
-            "{marker} {}",
-            pad_to_width(mode.label(), label_width.saturating_sub(2) as usize)
-        );
-        let text = pad_to_width(
-            &format!("{}{}", " ".repeat(leading), text),
-            area.width as usize,
-        );
-        let role = if focused {
-            R::SettingsValueFocused
-        } else {
-            R::SettingsValue
-        };
-        frame.render_widget(
-            Paragraph::new(Line::from(Span::styled(text, theme.style(role)))),
-            Rect {
-                x: area.x,
-                y,
-                width: area.width,
-                height: 1,
-            },
-        );
-        app.register_mouse_button(
-            Rect {
-                x: area.x,
-                y,
-                width: area.width,
-                height: 1,
-            },
-            MouseTarget::SettingsSpotifyImportModeSelect(mode),
-        );
-    }
-}
-
-pub(super) fn render_spotify_import_mode_dropdown_popup(frame: &mut Frame, app: &App, area: Rect) {
-    let Some(st) = app.settings.as_deref() else {
-        return;
-    };
-    let Some(selected) = st.spotify_import_mode_dropdown else {
-        return;
-    };
-    let popup =
-        crate::ui::centered_list_popup(area, crate::config::SpotifyImportMode::ALL.len(), 3, 32);
-    if popup.is_empty() {
-        return;
-    }
-
-    crate::ui::render_popup_background(frame, app, popup);
-    let block = Block::default()
-        .title(t!(" Spotify import mode ", " Spotify 가져오기 모드 "))
-        .borders(Borders::ALL)
-        .border_style(crate::ui::popup_style(app, R::Accent).add_modifier(Modifier::BOLD))
-        .style(crate::ui::popup_style(app, R::TextPrimary));
-    let inner = block.inner(popup);
-    frame.render_widget(block, popup);
-    if inner.height < 2 || inner.width < 8 {
-        crate::ui::seal_popup_background(frame, app, popup);
-        crate::ui::mark_art_rows_for_popup(frame, app, popup);
-        return;
-    }
-
-    let sections = Layout::vertical([Constraint::Min(1), Constraint::Length(1)]).split(inner);
-    let list_area = sections[0];
-    let row_w = list_area.width as usize;
-    let visible = list_area.height as usize;
-    let start = selected.saturating_sub(visible.saturating_sub(1));
-    for (idx, mode) in crate::config::SpotifyImportMode::ALL
-        .iter()
-        .copied()
-        .enumerate()
-        .skip(start)
-        .take(list_area.height as usize)
-    {
-        let focused = idx == selected;
-        let marker = if focused { "▶ " } else { "  " };
-        let text = pad_to_width(
-            &format!("{marker}{}", mode.label()),
-            row_w.saturating_sub(1),
-        );
-        let style = if focused {
-            crate::ui::anim::selection_style(
-                app,
-                Style::default()
-                    .fg(app.theme.color(R::SelectionFg))
-                    .bg(app.theme.color(R::SelectionBg))
-                    .add_modifier(Modifier::BOLD),
-            )
-        } else {
-            crate::ui::popup_style(app, R::TextPrimary)
-        };
-        let row = Rect {
-            x: list_area.x,
-            y: list_area.y + (idx - start) as u16,
-            width: list_area.width,
-            height: 1,
-        };
-        frame.render_widget(Paragraph::new(Line::from(text).style(style)), row);
-        app.register_mouse_button(row, MouseTarget::SettingsSpotifyImportModeSelect(mode));
-    }
-    frame.render_widget(
-        Paragraph::new(t!(
-            "↑↓ choose · Enter save · Esc close",
-            "↑↓ 선택 · Enter 저장 · Esc 닫기"
-        ))
-        .alignment(Alignment::Center)
-        .style(crate::ui::popup_style(app, R::TextMuted)),
-        sections[1],
-    );
-    crate::ui::seal_popup_background(frame, app, popup);
-    crate::ui::mark_art_rows_for_popup(frame, app, popup);
-}
-
 /// One field row: a left-aligned label and its current value (with a slider bar for numeric
 /// fields, `< … >` arrows for cycles, and a caret for the text field being edited).
 /// `display_row` is the row's position in the rendered list, used by the cascade reveal —
@@ -1078,518 +896,6 @@ fn field_row<'a>(
         Span::styled(label, fade(theme.style(R::SettingsLabel))),
         Span::styled(value, fade(theme.style(value_role))),
     ]))
-}
-
-/// A modal warning shown when a rebind on the Keys tab collides with an existing binding.
-/// Names the offending chord, the action already holding it, and where it lives, so the
-/// failure is loud and actionable rather than a silently-dropped remap.
-pub fn render_conflict(frame: &mut Frame, app: &App, area: Rect, conflict: &Conflict) {
-    let popup = centered_fixed(area, 54, 9);
-    crate::ui::render_popup_background(frame, app, popup);
-
-    let block = Block::default()
-        .title(t!(" ⚠ Keybinding conflict ", " ⚠ 단축키 충돌 "))
-        .borders(Borders::ALL)
-        .border_style(crate::ui::popup_style(app, R::Warning).add_modifier(Modifier::BOLD))
-        .style(crate::ui::popup_style(app, R::TextPrimary));
-    let inner = block.inner(popup);
-    frame.render_widget(block, popup);
-
-    let chord = keymap::format_chord_for_display(conflict.chord, app.retro_mode());
-    let where_line = if crate::i18n::is_korean() {
-        format!("{} 화면", conflict.ctx.title())
-    } else {
-        format!("in {}", conflict.ctx.title())
-    };
-    let lines = vec![
-        Line::from(""),
-        Line::from(vec![
-            Span::styled(
-                chord,
-                crate::ui::popup_style(app, R::Warning).add_modifier(Modifier::BOLD),
-            ),
-            Span::raw(t!(" is already bound to", " 은(는) 이미 사용 중")),
-        ]),
-        Line::from(Span::styled(
-            format!(
-                "\u{201c}{}\u{201d}",
-                conflict.existing.human_label_for(conflict.ctx)
-            ),
-            crate::ui::popup_style(app, R::Accent).add_modifier(Modifier::BOLD),
-        )),
-        Line::from(Span::styled(
-            where_line,
-            crate::ui::popup_style(app, R::TextMuted),
-        )),
-        Line::from(""),
-        Line::from(Span::styled(
-            t!(
-                "Binding unchanged · press any key",
-                "단축키 변경 안 됨 · 아무 키나 누르세요"
-            ),
-            crate::ui::popup_style(app, R::TextMuted),
-        )),
-    ];
-    frame.render_widget(
-        Paragraph::new(lines)
-            .alignment(Alignment::Center)
-            .style(crate::ui::popup_style(app, R::TextPrimary)),
-        inner,
-    );
-    crate::ui::seal_popup_background(frame, app, popup);
-    crate::ui::mark_art_rows_for_popup(frame, app, popup);
-}
-
-/// A modal confirmation for settings actions that have broad side effects.
-pub fn render_confirm(frame: &mut Frame, app: &App, area: Rect, confirm: SettingsConfirm) {
-    let popup = centered_fixed(area, 56, 9);
-    crate::ui::render_popup_background(frame, app, popup);
-
-    let block = Block::default()
-        .title(confirm.title())
-        .borders(Borders::ALL)
-        .border_style(crate::ui::confirm_border_style(app))
-        .style(crate::ui::popup_style(app, R::TextPrimary));
-    let inner = block.inner(popup);
-    frame.render_widget(block, popup);
-
-    let rows = Layout::vertical([
-        Constraint::Length(1),
-        Constraint::Length(1),
-        Constraint::Length(1),
-        Constraint::Length(1),
-        Constraint::Min(1),
-    ])
-    .split(inner);
-    let enabling_retro = app.settings.as_ref().is_some_and(|s| !s.draft.retro_mode);
-    frame.render_widget(
-        Paragraph::new(confirm.prompt(enabling_retro))
-            .alignment(Alignment::Center)
-            .style(crate::ui::popup_style(app, R::TextPrimary)),
-        rows[1],
-    );
-    frame.render_widget(
-        Paragraph::new(confirm.detail(enabling_retro))
-            .alignment(Alignment::Center)
-            .style(crate::ui::popup_style(app, R::TextMuted)),
-        rows[2],
-    );
-
-    let segs = [
-        buttons::Seg::button(
-            MouseTarget::ConfirmSettings,
-            t!(" Confirm (Enter) ", " 확인 (Enter) "),
-        ),
-        buttons::Seg::label("    "),
-        buttons::Seg::button(
-            MouseTarget::CancelSettings,
-            t!(" Cancel (Esc) ", " 취소 (Esc) "),
-        ),
-    ];
-    buttons::render_segments(
-        frame,
-        app,
-        rows[4],
-        &segs,
-        crate::ui::confirm_button_style(app),
-        crate::ui::confirm_gap_style(app),
-        Alignment::Center,
-    );
-    crate::ui::seal_popup_background(frame, app, popup);
-    crate::ui::mark_art_rows_for_popup(frame, app, popup);
-}
-
-/// The Spotify playlist picker overlay (Import from Spotify…): ↑/↓ select, Enter
-/// imports into a local Library playlist, Esc closes.
-pub fn render_spotify_picker(frame: &mut Frame, app: &App, area: Rect) {
-    let Some(picker) = app.overlays.spotify_picker.as_ref() else {
-        return;
-    };
-    let h = (picker.items.len() as u16 + 6).clamp(9, 20);
-    let popup = centered_fixed(area, 72, h);
-    crate::ui::render_popup_background(frame, app, popup);
-
-    let block = Block::default()
-        .title(t!(" Import from Spotify ", " Spotify에서 가져오기 "))
-        .borders(Borders::ALL)
-        .border_style(crate::ui::confirm_border_style(app))
-        .style(crate::ui::popup_style(app, R::TextPrimary));
-    let inner = block.inner(popup);
-    frame.render_widget(block, popup);
-
-    let rows = Layout::vertical([
-        Constraint::Length(2),
-        Constraint::Min(1),
-        Constraint::Length(1),
-    ])
-    .split(inner);
-    let import_mode = app
-        .settings
-        .as_ref()
-        .map(|st| st.draft.spotify_import_mode)
-        .unwrap_or(app.config.spotify.import_mode);
-    frame.render_widget(
-        Paragraph::new(vec![
-            Line::from(t!(
-                "Destination: Library > Playlists",
-                "대상: Library > Playlists"
-            )),
-            Line::from(format!(
-                "{}: {} · {}",
-                t!("Mode", "모드"),
-                import_mode.label(),
-                t!(
-                    "change in Settings > Accounts",
-                    "Settings > Accounts에서 변경"
-                )
-            )),
-        ])
-        .style(crate::ui::popup_style(app, R::TextMuted)),
-        rows[0],
-    );
-    let visible = rows[1].height as usize;
-    // Keep the selection in view for long playlist lists.
-    let first = picker
-        .selected
-        .saturating_sub(visible.saturating_sub(1))
-        .min(picker.items.len().saturating_sub(visible.max(1)));
-    let lines: Vec<ratatui::text::Line> = picker
-        .items
-        .iter()
-        .enumerate()
-        .skip(first)
-        .take(visible)
-        .map(|(i, item)| {
-            let marker = if i == picker.selected { "▸ " } else { "  " };
-            let count = if item.total > 0 {
-                format!("  ({})", item.total)
-            } else {
-                String::new()
-            };
-            let style = if i == picker.selected {
-                crate::ui::popup_style(app, R::TextPrimary)
-                    .add_modifier(ratatui::style::Modifier::BOLD)
-            } else {
-                crate::ui::popup_style(app, R::TextMuted)
-            };
-            ratatui::text::Line::styled(format!("{marker}{}{count}", item.label), style)
-        })
-        .collect();
-    frame.render_widget(Paragraph::new(lines), rows[1]);
-    // One click target per visible row: click selects, clicking the selected row (or a
-    // double-click) imports. Indices match the `.skip(first).take(visible)` render above.
-    for (vis, i) in (first..picker.items.len()).take(visible).enumerate() {
-        app.register_mouse_button(
-            Rect {
-                x: rows[1].x,
-                y: rows[1].y + vis as u16,
-                width: rows[1].width,
-                height: 1,
-            },
-            MouseTarget::SpotifyPickRow(i),
-        );
-    }
-    frame.render_widget(
-        Paragraph::new(t!(
-            "↑/↓ select · Enter import · Esc close · or click",
-            "↑/↓ 선택 · Enter 가져오기 · Esc 닫기 · 클릭 가능"
-        ))
-        .alignment(Alignment::Center)
-        .style(crate::ui::popup_style(app, R::TextMuted)),
-        rows[2],
-    );
-    crate::ui::seal_popup_background(frame, app, popup);
-    crate::ui::mark_art_rows_for_popup(frame, app, popup);
-}
-
-/// The radio-recording settings popup (opened from the Playback tab's one radio item). Rows:
-/// mode · min · max · output folder · keep-recent · notifications · browse.
-pub fn render_recording_settings(frame: &mut Frame, app: &App, area: Rect) {
-    use crate::config::{
-        RECORDING_MAX_SECONDS_MAX, RECORDING_MAX_SECONDS_MIN, RECORDING_MIN_SECONDS_MAX,
-        RECORDING_MIN_SECONDS_MIN, RECORDING_PAST_TRACKS_MAX, RECORDING_PAST_TRACKS_MIN,
-    };
-    let Some(popup) = app.overlays.recording_settings.as_ref() else {
-        return;
-    };
-    let Some(st) = app.settings.as_deref() else {
-        return;
-    };
-    let d = &st.draft;
-    let popup_rect = centered_fixed(area, 60, 13);
-    crate::ui::render_popup_background(frame, app, popup_rect);
-
-    let block = Block::default()
-        .title(t!(" Radio recording ", " 라디오 녹음 "))
-        .borders(Borders::ALL)
-        .border_style(crate::ui::confirm_border_style(app))
-        .style(crate::ui::popup_style(app, R::TextPrimary));
-    let inner = block.inner(popup_rect);
-    frame.render_widget(block, popup_rect);
-    let rows = Layout::vertical([Constraint::Min(1), Constraint::Length(1)]).split(inner);
-
-    let label_w = 16usize;
-    let value_w = usize::from(rows[0].width).saturating_sub(label_w + 2);
-    let dir_display = if popup.editing_dir {
-        crate::ui::text::editable_value(
-            &d.recording_dir,
-            popup.dir_cursor.byte_index(&d.recording_dir),
-            value_w,
-            crate::ui::anim::caret_char(app),
-            false,
-        )
-    } else if d.recording_dir.trim().is_empty() {
-        t!("(default)", "(기본값)").to_owned()
-    } else {
-        d.recording_dir.clone()
-    };
-    let entries: Vec<(String, String)> = vec![
-        (
-            t!("Mode", "모드").to_owned(),
-            format!("< {} >", d.recording_mode.label()),
-        ),
-        (
-            t!("Min duration", "최소 길이").to_owned(),
-            slider_str(
-                &bar(
-                    d.recording_min_seconds as f64,
-                    RECORDING_MIN_SECONDS_MIN as f64,
-                    RECORDING_MIN_SECONDS_MAX as f64,
-                ),
-                &format!("{}s", d.recording_min_seconds),
-            ),
-        ),
-        (
-            t!("Max duration", "최대 길이").to_owned(),
-            slider_str(
-                &bar(
-                    d.recording_max_seconds as f64,
-                    RECORDING_MAX_SECONDS_MIN as f64,
-                    RECORDING_MAX_SECONDS_MAX as f64,
-                ),
-                &format!("{} min", d.recording_max_seconds / 60),
-            ),
-        ),
-        (t!("Output folder", "저장 폴더").to_owned(), dir_display),
-        (
-            t!("Keep recent", "최근 보관").to_owned(),
-            slider_str(
-                &bar(
-                    d.recording_past_tracks as f64,
-                    RECORDING_PAST_TRACKS_MIN as f64,
-                    RECORDING_PAST_TRACKS_MAX as f64,
-                ),
-                &format!("{}", d.recording_past_tracks),
-            ),
-        ),
-        (
-            t!("Notifications", "알림").to_owned(),
-            if d.recording_notify {
-                "[x]".to_owned()
-            } else {
-                "[ ]".to_owned()
-            },
-        ),
-        (
-            t!("Browse recordings…", "녹음 목록 보기…").to_owned(),
-            "\u{21b5}".to_owned(),
-        ),
-    ];
-
-    let lines: Vec<Line> = entries
-        .iter()
-        .enumerate()
-        .map(|(i, (label, value))| {
-            let marker = if i == popup.row { "\u{25b8} " } else { "  " };
-            let style = if i == popup.row {
-                crate::ui::popup_style(app, R::TextPrimary).add_modifier(Modifier::BOLD)
-            } else {
-                crate::ui::popup_style(app, R::TextMuted)
-            };
-            Line::styled(
-                format!("{marker}{}{value}", pad_to_width(label, label_w)),
-                style,
-            )
-        })
-        .collect();
-    frame.render_widget(Paragraph::new(lines), rows[0]);
-    // Publish whole-row targets first; finer arrow/slider targets registered later win hit tests.
-    popup.rect.set(Some(popup_rect));
-    let right = rows[0].right();
-    for (i, (label, value)) in entries.iter().enumerate() {
-        let y = rows[0].y + i as u16;
-        if y >= rows[0].bottom() {
-            break;
-        }
-        app.register_mouse_button(
-            Rect {
-                x: rows[0].x,
-                y,
-                width: rows[0].width,
-                height: 1,
-            },
-            MouseTarget::RecordingRow(i),
-        );
-        // Measure the same display-width-padded prefix that was rendered above.
-        let marker = if i == popup.row { "\u{25b8} " } else { "  " };
-        let vx = rows[0].x
-            + buttons::text_width(marker)
-            + buttons::text_width(&pad_to_width(label, label_w));
-        let vw = buttons::text_width(value);
-        let put = |x: u16, w: u16, target: MouseTarget| {
-            if w == 0 || x >= right {
-                return;
-            }
-            app.register_mouse_button(
-                Rect {
-                    x,
-                    y,
-                    width: w.min(right - x),
-                    height: 1,
-                },
-                target,
-            );
-        };
-        // The adjustable rows expose arrow targets; numeric rows also expose their bar.
-        if matches!(i, 0 | 1 | 2 | 4) {
-            put(vx, 1, MouseTarget::RecordingChange { row: i, delta: -1 });
-            let last = vx.saturating_add(vw.saturating_sub(1));
-            put(last, 1, MouseTarget::RecordingChange { row: i, delta: 1 });
-            if i != 0 {
-                put(vx + 2, 11, MouseTarget::RecordingSlider(i));
-            }
-        }
-    }
-
-    let help = if popup.editing_dir {
-        t!(
-            "Type a folder \u{b7} Enter done \u{b7} Esc cancel",
-            "폴더 입력 \u{b7} Enter 완료 \u{b7} Esc 취소"
-        )
-    } else {
-        t!(
-            "\u{2191}/\u{2193} \u{b7} \u{2190}/\u{2192} change \u{b7} Enter \u{b7} Esc close",
-            "\u{2191}/\u{2193} \u{b7} \u{2190}/\u{2192} 변경 \u{b7} Enter \u{b7} Esc 닫기"
-        )
-    };
-    frame.render_widget(
-        Paragraph::new(help)
-            .alignment(Alignment::Center)
-            .style(crate::ui::popup_style(app, R::TextMuted)),
-        rows[1],
-    );
-    crate::ui::seal_popup_background(frame, app, popup_rect);
-    crate::ui::mark_art_rows_for_popup(frame, app, popup_rect);
-}
-
-/// The recordings browser: the in-progress track (if any) plus recent finished tracks, with
-/// per-row save/discard/play. Reads `app.recorder`.
-pub fn render_recordings_browser(frame: &mut Frame, app: &App, area: Rect) {
-    let Some(browser) = app.overlays.recordings_browser.as_ref() else {
-        return;
-    };
-    let mut items: Vec<(String, String)> = Vec::new();
-    if let Some(seg) = app.recorder.current.as_ref() {
-        items.push((
-            format!("\u{25cf} {}", seg_display(seg)),
-            t!("Recording\u{2026}", "녹음 중\u{2026}").to_owned(),
-        ));
-    }
-    for t in app.recorder.history.iter() {
-        let glyph = if matches!(t.state, crate::recorder::RecordingState::Saved) {
-            "\u{25a3}"
-        } else if t.state.is_recorded() {
-            "\u{2713}"
-        } else {
-            "\u{b7}"
-        };
-        items.push((
-            format!("{glyph} {}", t.display()),
-            format!("{}  {}", fmt_dur(t.duration_secs), t.state.label()),
-        ));
-    }
-
-    let n = items.len();
-    let h = ((n as u16) + 4).clamp(7, 20);
-    let popup = centered_fixed(area, 64, h);
-    // Publish the browser rect so a click outside closes it (mirrors the queue window).
-    browser.rect.set(Some(popup));
-    crate::ui::render_popup_background(frame, app, popup);
-    let block = Block::default()
-        .title(t!(" Radio recordings ", " 라디오 녹음 목록 "))
-        .borders(Borders::ALL)
-        .border_style(crate::ui::confirm_border_style(app))
-        .style(crate::ui::popup_style(app, R::TextPrimary));
-    let inner = block.inner(popup);
-    frame.render_widget(block, popup);
-    let rows = Layout::vertical([Constraint::Min(1), Constraint::Length(1)]).split(inner);
-    let visible = rows[0].height as usize;
-
-    if items.is_empty() {
-        frame.render_widget(
-            Paragraph::new(t!("No recordings yet.", "아직 녹음이 없어요."))
-                .alignment(Alignment::Center)
-                .style(crate::ui::popup_style(app, R::TextMuted)),
-            rows[0],
-        );
-    } else {
-        let sel = browser.selected.min(n.saturating_sub(1));
-        let first = sel
-            .saturating_sub(visible.saturating_sub(1))
-            .min(n.saturating_sub(visible.max(1)));
-        let lines: Vec<Line> = items
-            .iter()
-            .enumerate()
-            .skip(first)
-            .take(visible)
-            .map(|(i, (title, sub))| {
-                let marker = if i == sel { "\u{25b8} " } else { "  " };
-                let style = if i == sel {
-                    crate::ui::popup_style(app, R::TextPrimary).add_modifier(Modifier::BOLD)
-                } else {
-                    crate::ui::popup_style(app, R::TextMuted)
-                };
-                Line::styled(format!("{marker}{title}   {sub}"), style)
-            })
-            .collect();
-        frame.render_widget(Paragraph::new(lines), rows[0]);
-        // One clickable rect per visible row → single-click selects that item.
-        for (vis_idx, i) in (first..n).take(visible).enumerate() {
-            app.register_mouse_button(
-                Rect {
-                    x: rows[0].x,
-                    y: rows[0].y + vis_idx as u16,
-                    width: rows[0].width,
-                    height: 1,
-                },
-                MouseTarget::RecordingBrowseRow(i),
-            );
-        }
-    }
-    frame.render_widget(
-        Paragraph::new(t!(
-            "\u{2191}/\u{2193} \u{b7} s save \u{b7} d discard \u{b7} Enter play \u{b7} Esc close",
-            "\u{2191}/\u{2193} \u{b7} s 저장 \u{b7} d 삭제 \u{b7} Enter 재생 \u{b7} Esc 닫기"
-        ))
-        .alignment(Alignment::Center)
-        .style(crate::ui::popup_style(app, R::TextMuted)),
-        rows[1],
-    );
-    crate::ui::seal_popup_background(frame, app, popup);
-    crate::ui::mark_art_rows_for_popup(frame, app, popup);
-}
-
-fn seg_display(seg: &crate::recorder::OpenSegment) -> String {
-    let base = seg.raw.trim();
-    if base.is_empty() {
-        seg.title.clone().unwrap_or_else(|| "\u{2014}".to_owned())
-    } else {
-        base.to_owned()
-    }
-}
-
-fn fmt_dur(secs: u32) -> String {
-    format!("{}:{:02}", secs / 60, secs % 60)
 }
 
 /// A `w`×`h` rect centered in `area`, clamped so it never exceeds the available space.
